+++ src/cruder/list.rs
@@ -1,7 +1,7 @@
 ///
 /// CRUD sur les données
 ///
-use sqlx::{Pool, Postgres, Sqlite};
+use sqlx::{Pool, Postgres, Sqlite, Error};
 
 use crate::lexicer::lex_application::Application;
 use crate::lexicer::lex_table::Element;
@@ -19,7 +19,7 @@ pub async fn crud_list(
     tableid: &str,
     viewid: &str,
     id: &str,
-) -> Vec<HashMap<String, Element>> {
+) -> Result<Vec<HashMap<String, Element>>, Error> {
     // construction de l'ordre sql
     let mut sql = "SELECT ".to_string();
     // on prend les colonnes définies dans la view.velements
@@ -79,8 +79,7 @@ pub async fn crud_list(
         &application,
         &view.velements,
         table,
-    )
-    .await;
-    records
+    ).await?;
+    Ok(records)
 }