+++ src/cruder/mod.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 
-use actix_web_flash_messages::FlashMessage;
-use sqlx::{Postgres, Pool, Sqlite};
+use sqlx::{Postgres, Pool, Sqlite, Error};
 
 use crate::lexicer::{lex_application::Application, lex_table::{Element, Table}};
 
@@ -23,7 +22,7 @@ pub async fn records_elements(
     application: &Application,
     velements: &Vec<Element>,
     table: &Table,
-) -> Vec<HashMap<String, Element>> {
+) -> Result<Vec<HashMap<String, Element>>, Error> {
 
     let mut records = Vec::new();
     let mut vrows: Vec<HashMap<String, String>> = Vec::new();
@@ -35,13 +34,7 @@ pub async fn records_elements(
             vrows.push(hvalue);
         }
     } else {
-        let rows = match sqlx::query(&sql).fetch_all(pooldb).await {
-            Ok(t) => t,
-            Err(e) => {
-                FlashMessage::info(format!("{:?}", e)).send();
-                Vec::new()
-            }
-        };
+        let rows = sqlx::query(&sql).fetch_all(pooldb).await?;
         // Chargement des enregistrements dans un tableau de valeur
         vrows = rows_to_vmap(rows);
     }
@@ -59,7 +52,7 @@ pub async fn records_elements(
         // TODO
         for vel in velements {
             let mut element = vel.clone();
-            element.compute_value(poolite, &hvalue).await;
+            element.compute_value(poolite, &hvalue).await?;
             element.key_value = key_value.clone();
             hel.insert(vel.elid.clone(), element);
         }
@@ -76,7 +69,7 @@ pub async fn records_elements(
         // calcul des autres propriétés
         for vel in velements {
             let mut element = hel.get(&vel.elid).unwrap().clone();
-            element.compute_prop(pooldb, poolite, &hvalue_computed).await;
+            element.compute_prop(pooldb, poolite, &hvalue_computed).await?;
             if vel.elid == table.setting.key {
                 element.read_only = true;
             }
@@ -85,6 +78,6 @@ pub async fn records_elements(
         records.push(hel);
     }
 
-    records
+    Ok(records)
 }