+++ src/middler/mod.rs
@@ -1,0 +1,9 @@
+
+
+pub mod mid_session;
+pub mod flash;
+pub use self::flash::set_flash;
+pub use self::flash::get_flash;
+pub use self::flash::clear_flash;
+// mod history;
+// pub use self::history::History;