+++ src/router/route_add_post.rs
@@ -1,11 +1,13 @@
 //! Ouverture d'une view
 //!
+use crate::middler::set_flash;
 use crate::AppState;
 use crate::cruder::update::crud_update;
+use crate::middler::flash::FlashMessage;
+use actix_session::Session;
 use actix_web::{HttpResponse, web};
 use actix_web::http::header::LOCATION;
 use actix_web::web::Path;
-use actix_web_flash_messages::IncomingFlashMessages;
 use std::{
     // collections::HashMap,
     sync::atomic::Ordering
@@ -16,7 +18,7 @@ pub async fn add_post(
     path: Path<(String, String, String, String, String)>,
     web::Form(form_posted): web::Form<Vec<(String, String)>>,
     data: web::Data<AppState>,
-    _flash: IncomingFlashMessages
+    session: Session,
 ) -> HttpResponse {
 
 
@@ -28,17 +30,20 @@ pub async fn add_post(
     // let view = table.views.get(&viewid).unwrap();
     let form = table.forms.get(&formid).unwrap();
 
-    let result = crud_update(&data.db, &data.dblite,
-        &table, &form.velements, &id, &form_posted).await;
+    let mut location = String::new();
+    match crud_update(&data.db, &data.dblite, &table, &form.velements, &id, &form_posted).await {
+            Ok(s) => {
+                set_flash(&session, FlashMessage::success(&s)).unwrap();
+                location.push_str(format!("/form/{appid}/{tableid}/{viewid}/{formid}/{id}").as_str());
+            },
+            Err(e) => {
+                set_flash(&session, FlashMessage::success(format!("{e:?}").as_str())).unwrap();
+                location.push_str(format!("/edit/{appid}/{tableid}/{viewid}/{formid}/{id}").as_str());
+            }
+        };
+        HttpResponse::SeeOther()
+        .insert_header((LOCATION, location))
+        .finish()
 
-    if result {
-        HttpResponse::SeeOther()
-        .insert_header((LOCATION, format!("/view/{appid}/{tableid}/{viewid}/{formid}/{id}")))
-        .finish()
-    } else {
-        HttpResponse::SeeOther()
-        .insert_header((LOCATION, format!("/edit/{appid}/{tableid}/{viewid}/{formid}/{id}")))
-        .finish()
-    }
 }