+++ src/router/route_edit_post.rs
@@ -2,10 +2,12 @@
 //!
 use crate::AppState;
 use crate::cruder::update::crud_update;
+use crate::middler::flash::FlashMessage;
+use crate::middler::set_flash;
+use actix_session::Session;
 use actix_web::{HttpResponse, web};
 use actix_web::http::header::LOCATION;
 use actix_web::web::Path;
-use actix_web_flash_messages::{FlashMessage, IncomingFlashMessages};
 use std::{
     // collections::HashMap,
     sync::atomic::Ordering
@@ -15,7 +17,7 @@ pub async fn edit_post(
     path: Path<(String, String, String, String, String)>,
     web::Form(form_posted): web::Form<Vec<(String, String)>>,
     data: web::Data<AppState>,
-    _flash: IncomingFlashMessages
+    session: Session,
 ) -> HttpResponse {
 
     let (appid, tableid, viewid, formid, id) = path.into_inner();
@@ -26,19 +28,19 @@ pub async fn edit_post(
     // let view = table.views.get(&viewid).unwrap();
     let form = table.forms.get(&formid).unwrap();
 
-    let result = crud_update(&data.db, &data.dblite, &table, &form.velements, &id, &form_posted).await;
-
-    FlashMessage::info("route_edit_post").send();
-    FlashMessage::error("ya un bug!!").send();
-
-    if result {
+    let mut location = String::new();
+    match crud_update(&data.db, &data.dblite, &table, &form.velements, &id, &form_posted).await {
+            Ok(s) => {
+                set_flash(&session, FlashMessage::success(&s)).unwrap();
+                location.push_str(format!("/form/{appid}/{tableid}/{viewid}/{formid}/{id}").as_str());
+            },
+            Err(e) => {
+                set_flash(&session, FlashMessage::success(format!("{e:?}").as_str())).unwrap();
+                location.push_str(format!("/edit/{appid}/{tableid}/{viewid}/{formid}/{id}").as_str());
+            }
+        };
         HttpResponse::SeeOther()
-        .insert_header((LOCATION, format!("/form/{appid}/{tableid}/{viewid}/{formid}/{id}")))
+        .insert_header((LOCATION, location))
         .finish()
-    } else {
-        HttpResponse::SeeOther()
-        .insert_header((LOCATION, format!("/edit/{appid}/{tableid}/{viewid}/{formid}/{id}")))
-        .finish()
-    }
 }