+++ src/service/mod.rs
@@ -1,6 +1,0 @@
-
-
-pub mod srv_session;
-
-// mod history;
-// pub use self::history::History;